extract bulkdatatransfer
